Update test checking null severity (#277)
The new test in HealthStatusTest that checks that the HealthStatus
is WARN when given a null severity should test when "healthy" is
true or false. We want the returned HealthStatus to be WARN
anytime we find a "severity" detail with a null value, since it
indicates a problem somewhere else, i.e., in the health check
that set the null severity detail.